charity_project/charity/requirements.txt
argon2-cffi==20.1.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
Django==2.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.11
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
wincertstore==0.2
